Moved getUrlFromList to filterUtil.
