+++ src/Filter/modules/MultiSelectFilter/__tests__/multiSelectUtil.test.js
@@ -1,4 +1,4 @@
-import { createOption, getUrlFromList } from '../multiSelectUtil';
+import { createOption } from '../multiSelectUtil';
 
 describe('createOption', () => {
     it('returns an object with the correct label and value.', () => {
@@ -14,30 +14,4 @@ describe('createOption', () => {
             value: undefined
         });
     });
-});
-
-describe('getUrlFromList', () => {
-    it('appends each list item to the url with correct name', () => {
-        expect(getUrlFromList('item', ['red', 'blue'], 3)).toBe('&item=red&item=blue')
-    });
-
-    describe('when list is an array', () => {
-        it('returns an empty string if list size matches count.', () => {
-            expect(getUrlFromList('item', ['red', 'blue'], 2)).toBe('')
-        });
-
-        it('returns an empty string if list size is 0.', () => {
-            expect(getUrlFromList('item', [], 2)).toBe('')
-        });
-    });
-
-    describe('when list is a set', () => {
-        it('returns an empty string if list size matches count.', () => {
-            expect(getUrlFromList('item', new Set(['red', 'blue']), 2)).toBe('')
-        });
-
-        it('returns an empty string if list size is 0.', () => {
-            expect(getUrlFromList('item', new Set(), 2)).toBe('')
-        });
-    })
 });
