+++ src/Filter/util/filterUtil.ts
@@ -46,3 +46,23 @@ export const getFilterAppliedCount = (filters: FilterSet, filterValues: FilterVa
     });
     return count;
 };
+
+/**
+ * Given a list (Array or Set) of string, return a query string representation.
+ * Default behavior is that when all items are selected, nothing is returned.
+ */
+ export const getUrlFromList = (name: string, list: string[] | Set<string>, count: number) => {
+    const listSize = Array.isArray(list) ? list.length : list.size;
+
+    // if all items in list selected, pass empty string in url filter
+    if (listSize === count || listSize === 0) {
+        return '';
+    }
+
+    const url = new URLSearchParams();
+
+    for (const item of list) {
+        url.append(name, item);
+    }
+    return `&${url}`;
+};
